split identities a bit more
