+++ service/src/db/mod.rs
@@ -9,6 +9,10 @@ mod identity_error;
 pub use self::identity_error::*;
 mod identity_identities;
 pub use self::identity_identities::*;
+mod identity_external_link;
+pub use self::identity_external_link::*;
+mod identity_search;
+pub use self::identity_search::*;
 mod identity_uow;
 pub use self::identity_uow::*;
 mod identity_roles;
